Add files via upload
1. The majority of the film was made in March, with an annual average of roughly 2300 films.
2. Drama is the most popular genre with 4285 films averaging 106.7746 minutes in length. 
RSVP movies can concentrate on drama, thriller, and action genres.
3. Most films received a score of 6-8. Aiming for a median rating of 8+ will 
boost the likelihood of a super successful film.
4. Dream Warriors Pictures and National Theatre Live have produced films that have 
received the most outstanding ratings in the business. With a large number of multilingual films and an 
emphasis on Indian audiences, Star Cinema and Twentieth Century Fox are also strong 
contenders.
5. James Mangold, Anthony Russo, Joe Russo, and Soubin Shahir are among the top 
directors identified by the study. The last three are tied for second place. RSVP can work 
with these directors on future projects to increase hit pictures.
6. Mammooty and Mohanlal are the top actors with the most ratings in India.
7. Taapsee Pannu may be considered for future projects because she is at the top of the list 
with a 7.74 average rating.
8. Marvel Studios, Twentieth Century Fox, and Warner Bros can be chosen as Global Partners 
because their films have received the most votes.
9. Vijay Sethupati can focus on India because he is well-known for his films.
10. The top two production firms that have generated the most hits among multilingual 
movies are Star Cinema and Twentieth Century Fox.